Moved tests to src folder.
