+++ pwa/src/__tests__/database/create_stock_info.spec.ts
@@ -6,9 +6,9 @@ import {
   Repository,
 } from 'typeorm'
 
-import { migrations } from '../../src/data/migrations'
-import { StockInfoModel } from '../../src/data/entities/StockInfoModel'
-import { TransactionModel } from '../../src/data/entities/TransactionModel'
+import { migrations } from '../../data/migrations'
+import { StockInfoModel } from '../../data/entities/StockInfoModel'
+import { TransactionModel } from '../../data/entities/TransactionModel'
 
 describe('@Database: Create stock info', () => {
   const pathToDatabase = path.resolve(__dirname, 'test_database.db')
